backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Selenium-based web scraper for retrieving switch2 usage data"
requires-python = ">=3.12"
dependencies = [
    "selenium>=4.24.0",
    "requests-html>=0.10.0",
    "lxml-html-clean>=0.2.1",
]
